Fixes config parsing and stages
