+++ crates/primitives/src/components/layer_two.rs
@@ -1,4 +1,4 @@
-use std::{fmt::Display, str::FromStr};
+use std::fmt::Display;
 
 use enum_variants_strings::EnumVariantsStrings;
 use serde::{Deserialize, Serialize};
@@ -9,9 +9,7 @@ use strum::EnumIter;
 /// The OP Stack L2 client is a minimally modified version of the L1 client
 /// that supports deposit transactions as well as a few other small OP-specific
 /// changes.
-#[derive(
-    Default, Debug, Clone, PartialEq, EnumVariantsStrings, Deserialize, Serialize, EnumIter,
-)]
+#[derive(Default, Clone, PartialEq, EnumVariantsStrings, Deserialize, Serialize, EnumIter)]
 #[enum_variants_strings_transform(transform = "kebab_case")]
 pub enum L2Client {
     /// OP Geth
@@ -23,7 +21,13 @@ pub enum L2Client {
     OpReth,
 }
 
-impl FromStr for L2Client {
+impl std::fmt::Debug for L2Client {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.to_str())
+    }
+}
+
+impl std::str::FromStr for L2Client {
     type Err = eyre::Report;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
@@ -43,5 +47,32 @@ impl FromStr for L2Client {
 impl Display for L2Client {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.to_str())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_debug_string() {
+        assert_eq!(format!("{:?}", L2Client::OpGeth), "op-geth");
+        assert_eq!(format!("{:?}", L2Client::OpErigon), "op-erigon");
+        assert_eq!(format!("{:?}", L2Client::OpReth), "op-reth");
+    }
+
+    #[test]
+    fn test_l2_client_from_str() {
+        assert_eq!("op-geth".parse::<L2Client>().unwrap(), L2Client::OpGeth);
+        assert_eq!("op-erigon".parse::<L2Client>().unwrap(), L2Client::OpErigon);
+        assert_eq!("op-reth".parse::<L2Client>().unwrap(), L2Client::OpReth);
+        assert!("invalid".parse::<L2Client>().is_err());
+    }
+
+    #[test]
+    fn test_l2_client_display() {
+        assert_eq!(L2Client::OpGeth.to_string(), "op-geth");
+        assert_eq!(L2Client::OpErigon.to_string(), "op-erigon");
+        assert_eq!(L2Client::OpReth.to_string(), "op-reth");
     }
 }
